feat: DB health check API 패키지 변경 및 주석 추가
